Moved around interfaces some more
