+++ flare/client/bitcoin/rpc_client.go
@@ -1,0 +1,9 @@
+package bitcoin
+
+import (
+	"github.com/ybbus/jsonrpc/v2"
+)
+
+type RPCClient interface {
+	Call(method string, params ...interface{}) (*jsonrpc.RPCResponse, error)
+}
