+++ flare/connector/bitcoin/api_client.go
@@ -1,6 +1,6 @@
 package bitcoin
 
-type API interface {
+type APIClient interface {
 	Block(hash [32]byte) (*Block, error)
 	Transaction(hash [32]byte, index uint8) (*Transaction, error)
 }
