+++ flare/connector/ripple/api_client.go
@@ -1,6 +1,6 @@
 package ripple
 
-type API interface {
+type APIClient interface {
 	Ledger(index uint32) (*Ledger, error)
 	Transaction(hash [32]byte) (*Transaction, error)
 }
